moved filter to other package
